Rename app to main-container
