+++ lib/react/main-container.js
