+++ lib/session.js
@@ -11,7 +11,7 @@ import Disposables from "./disposables";
 import * as A from "./action-creators.js";
 import NotesFileFilter from "./notes-file-filter";
 import newRootReducer from "./reducers/index";
-import newApp from "./react/app";
+import mainContainer from "./react/main-container";
 import NotesCache from "./notes-cache";
 import RestartSessionForNewConfigToTakeEffect from "./restart-session-for-new-config-to-take-effect";
 import ToggleAtomWindow from "./toggle-atom-window";
@@ -87,10 +87,10 @@ export default class Session {
     };
     this._store = createStore(rootReducer, initialState, middlewares);
 
-    const App = newApp(columns);
+    const MainContainer = mainContainer(columns);
     render(
       <Provider store={this._store}>
-        <App />
+        <MainContainer />
       </Provider>,
       this._panel.getItem()
     );
